test. Complete testing of setup app so far
The redirect problems I had in the setup app is now solved. At this
point in the project the coverage report shows 100% tested.

However I still have som problems in the main app.

Since automated testing is not mandatory in the project and it takes a
lot of time from the rest of the project, I've decided to be content
with the testing I've done so far and be pleased with everything I've
learned from it.