style: Add styling to home page, belongs to #44
This commit adds styling to the home page according to the style on the
session page. A hover function over days has been added
to the calendar so that the whole cell is a link to session page.
If the day doesn't have a planned session the create session page will
be rendered instead. (This function will come in a later commit)