fix: Reverse changes in an earlier commit
Some of the changes made in commit
"style: Add styling to home page, belongs to #44" on may 12, had to
be reversed.
The code I planned to implement where the user could add a session by
clicking on a date will have to be postponed.
If there is still time later on in the project this feature will be
implemented.

The button to create a session was therefor reinstalled on the calendar
page and coding done connected to this feature was reversed.