fix: Trying to fix edit/create exercise
A previous commit introduced en error when editing an exercise.
This error is solved with this commit, but there is still an error left.
Muscle groups selected in a multichoicefield are not saved to the
database after the user filter is added.
I make this commit so my fellow slackers can fork the repo and have a go